PortfolioCraft/requirements.txt
pip install playwright
python -m playwright install chromium
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cssselect2==0.8.0
distro==1.9.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.96.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
playwright==1.55.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
pyparsing==3.2.3
pypdfium2==4.30.1
pyphen==0.17.2
python-docx==1.2.0
python-dotenv==1.1.1
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
